requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev.txt dev.in
#
astroid==1.5.2            # via edx-lint, pylint, pylint-celery
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click-log==0.1.8          # via edx-lint
click==6.7                # via click-log, edx-lint, pip-tools
diff-cover==1.0.4
django==1.11.14           # via edx-i18n-tools
edx-i18n-tools==0.4.6
edx-lint==0.5.5
first==2.0.1              # via pip-tools
idna==2.7                 # via requests
inflect==0.3.1            # via jinja2-pluralize
isort==4.3.4              # via pylint
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via diff-cover, jinja2-pluralize
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via pylint
packaging==17.1           # via tox
path.py==11.0.1           # via edx-i18n-tools
pip-tools==2.0.2
pkginfo==1.4.2            # via twine
pluggy==0.6.0             # via tox
polib==1.1.0              # via edx-i18n-tools
py==1.5.4                 # via tox
pygments==2.2.0           # via diff-cover
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.3  # via pylint-celery, pylint-django
pylint==1.7.1             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.2.0          # via packaging
pytz==2018.5              # via django
pyyaml==3.13              # via edx-i18n-tools
requests-toolbelt==0.8.0  # via twine
requests==2.19.1          # via requests-toolbelt, twine
six==1.11.0               # via astroid, diff-cover, edx-i18n-tools, edx-lint, packaging, pip-tools, pylint, tox
tox-battery==0.5.1
tox==3.1.1
tqdm==4.23.4              # via twine
twine==1.11.0
urllib3==1.23             # via requests
virtualenv==16.0.0        # via tox
wheel==0.31.1
wrapt==1.10.11            # via astroid
